Inserted inner space between two letters.
The new property itself only states that apart from the 'A' rows, each row
must have exactly two identical letters. Because the tests for the 'A'
rows, together with the tests for symmetric contours, imply that each row
must have a width of an uneven number, and again because of the symmetric
contour requirement, one has to introduce space between the two
characters. Using the Devil's Advocate technique, it seems that the
simplest way of passing all tests is to fill out the inner space
completely. Here's an example of calling Diamond.make 'D' with the current
implementation:

   A
B     B
C     C
D     D
C     C
B     B
   A