Made the bottom part a symmetric triangle.
The property specifies that the lower left 'corner' should be a triangle,
and due to the tests for symmetry over the vertical axis, the lower right
corner must be a triangle as well.

The implementation uses Devil's Advocate, and goes quite a bit out of its
way to make the top of the figure wrong. As you'll see shortly, it will
actually be a simpler implementation to keep the figure symmetric around
the horizontal axis as well, but we should have that as an explict
property.

The current implementation produces the figure below when Diamond.make is
called with 'D':

   A
  B B
  C C
D     D
 C   C
  B B
   A